Removing new_stand in setup_canopy
new_stand is already called in the constructor of Adel therefore in the first call
to setup_canopy it is not necessary. In calls to setup_canopy from
grow and grow_dd it would have the undesired effect that it re-samples
the plant positions, causing them to grow in different places that they
did before. This would not be realistic growth.